Add AlreadyResumedException and update effect handler functionality
A new exception type `AlreadyResumedException` is added for cases where the resume function in the `EffectHandler` interface is called more than once within a `with` block. This is due to a restriction in Kotlin where multi-shot continuations are not supported. The `invokeHandler` function's access level is also modified from `public` to `private`. Finally, new test scenarios for object `Next` effect are introduced in the `EffectsTesting` class.